requirements.txt
radon>=6.0.1
lizard>=1.17.13
pylint>=3.0.3
PyGithub>=2.1.1
python-dotenv>=1.0.1
anthropic>=0.39.0
jinja2>=3.1.3
PyYAML>=6.0.1
pytz>=2024.1
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
weasyprint>=61.0
cairocffi>=1.6.0
